requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jwt==1.3.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
